[#9] refactor(rename) :  - 용어 일관성을 위해 Notification -> Remind 수정  - Batch 관련된 코드는 수정 X
