--- src/main/kotlin/com/yapp/web2/domain/notification/repository/NotificationRepository.kt
@@ -1,7 +1,0 @@
-package com.yapp.web2.domain.notification.repository
-
-import org.springframework.stereotype.Repository
-
-@Repository
-interface NotificationRepository {
-}
