+++ src/main/kotlin/com/yapp/web2/domain/remind/controller/RemindController.kt
@@ -1,8 +1,8 @@
-package com.yapp.web2.domain.notification.controller
+package com.yapp.web2.domain.remind.controller
 
-import com.yapp.web2.domain.notification.entity.dto.RemindCycleRequest
-import com.yapp.web2.domain.notification.entity.dto.RemindToggleRequest
-import com.yapp.web2.domain.notification.service.NotificationService
+import com.yapp.web2.domain.remind.entity.dto.RemindCycleRequest
+import com.yapp.web2.domain.remind.entity.dto.RemindToggleRequest
+import com.yapp.web2.domain.remind.service.RemindService
 import com.yapp.web2.util.Message
 import io.swagger.annotations.ApiOperation
 import io.swagger.annotations.ApiParam
@@ -13,8 +13,8 @@ import javax.servlet.http.HttpServletRequest
 
 @RestController
 @RequestMapping("/api/v1/mypage/remind")
-class NotificationController(
-    private val notificationService: NotificationService
+class RemindController(
+    private val remindService: RemindService
 ) {
 
     @ApiOperation(value = "리마인드 알람 설정(토글) API")
@@ -24,7 +24,7 @@ class NotificationController(
         @RequestBody @ApiParam(value = "리마인드 토글(true / false)", required = true) request: RemindToggleRequest
     ): ResponseEntity<String> {
         val accessToken = servletRequest.getHeader("AccessToken")
-        notificationService.changeRemindAlarm(request, accessToken)
+        remindService.changeRemindAlarm(request, accessToken)
         return ResponseEntity.status(HttpStatus.OK).body(Message.SUCCESS)
     }
 
@@ -36,7 +36,7 @@ class NotificationController(
         @RequestBody @ApiParam(value = "리마인드 알람 주기 설정 정보", required = true) request: RemindCycleRequest
     ): ResponseEntity<String> {
         val accessToken = servletRequest.getHeader("AccessToken")
-        notificationService.updateRemindAlarmCycle(request, accessToken)
+        remindService.updateRemindAlarmCycle(request, accessToken)
         return ResponseEntity.status(HttpStatus.OK).body(Message.SUCCESS)
     }
