+++ src/main/kotlin/com/yapp/web2/domain/remind/entity/Remind.kt
@@ -1,4 +1,4 @@
-package com.yapp.web2.domain.notification.entity
+package com.yapp.web2.domain.remind.entity
 
 import com.yapp.web2.domain.BaseTimeEntity
 import com.yapp.web2.domain.account.entity.Account
@@ -6,7 +6,7 @@ import java.time.LocalDateTime
 import javax.persistence.*
 
 @Entity
-class Notification(
+class Remind(
     var remindSetTime: LocalDateTime,
     var remindCycle: LocalDateTime,
