+++ src/main/kotlin/com/yapp/web2/domain/remind/entity/dto/RemindToggleRequest.kt
@@ -1,4 +1,4 @@
-package com.yapp.web2.domain.notification.entity.dto
+package com.yapp.web2.domain.remind.entity.dto
 
 class RemindToggleRequest(
     val remindToggle: Boolean
