+++ src/main/kotlin/com/yapp/web2/domain/remind/repository/RemindRepository.kt
@@ -1,0 +1,7 @@
+package com.yapp.web2.domain.remind.repository
+
+import org.springframework.stereotype.Repository
+
+@Repository
+interface RemindRepository {
+}
